+++ graphistry/Plottable.py
@@ -17,6 +17,7 @@ from graphistry.privacy import Mode as PrivacyMode, Privacy, ModeAction
 from graphistry.Engine import EngineAbstractType
 from graphistry.utils.json import JSONVal
 from graphistry.client_session import ClientSession, AuthManagerProtocol
+from graphistry.models.types import ValidationParam
 
 if TYPE_CHECKING:
     try:
@@ -779,8 +780,16 @@ class Plottable(Protocol):
 
     def to_cudf(self) -> 'Plottable':
         ...
-    
+
     def to_pandas(self) -> 'Plottable':
+        ...
+
+    def to_arrow(
+        self,
+        table: Optional[Any] = None,
+        validate: ValidationParam = 'autofix',
+        warn: bool = True
+    ) -> Optional[Any]:
         ...
 
     def protocol(self, v: Optional[str] = None) -> str:
@@ -805,8 +814,9 @@ class Plottable(Protocol):
     def upload(
         self,
         memoize: bool = True,
-        erase_files_on_fail=True,
-        validate: bool = True
+        erase_files_on_fail: bool = True,
+        validate: ValidationParam = 'autofix',
+        warn: bool = True
     ) -> 'Plottable':
         ...
 
@@ -823,7 +833,8 @@ class Plottable(Protocol):
         erase_files_on_fail: bool = True,
         extra_html: str = "",
         override_html_style: Optional[str] = None,
-        validate: bool = True
+        validate: ValidationParam = 'autofix',
+        warn: bool = True
     ) -> Any:
         ...